+++ src/Bio/FASTA/Parser.hs
@@ -6,7 +6,7 @@ import           Bio.FASTA.Type         (Fasta, FastaItem(..))
 import           Bio.Sequence           (BareSequence, bareSequence)
 import           Data.Attoparsec.Text   (Parser, many', many1', char, endOfLine, letter,
                                             takeWhile, choice, endOfInput)
-import           Data.Text              (Text)
+import           Data.Text              (Text, strip)
 import           Prelude         hiding (takeWhile)
 
 -- | Parser of .fasta file.
@@ -18,10 +18,10 @@ item :: Parser (FastaItem Char)
 item = FastaItem <$> seqName <*> fastaSeq
 
 seqName :: Parser (Text)
-seqName = char '>' *> takeWhile (`notElem` ['\n', '\r', '\t']) <* eol
+seqName = strip <$> (char '>' *> tabs *> takeWhile (`notElem` ['\n', '\r']) <* tabs <* eol)
 
 fastaSeq :: Parser (BareSequence Char)
-fastaSeq = bareSequence . mconcat <$> many1' line
+fastaSeq = bareSequence . mconcat <$> many' line
 
 line :: Parser (String)
 line = many1' letter <* eol