+++ src/Bio/FASTA/Writer.hs
@@ -5,7 +5,7 @@ module Bio.FASTA.Writer
 import           Bio.FASTA.Type     (Fasta, FastaItem(..))
 import           Bio.Sequence       (BareSequence, sequ)
 import           Control.Lens       ((^.))
-import           Data.Text          (Text, pack, append)
+import           Data.Text          (Text, pack)
 import           Data.List.Split    (chunksOf)
 import           Data.Vector        (Vector, toList)
 import           Prelude     hiding (drop)